Reorganize tests: split ExecutionContext + measurement clusters
Per @elalish's suggestion on #1687 and review on #1697.

Two extractions from manifold_test.cpp:

- test/context_test.cpp: ExecutionContext and ForEachCtx tests, plus
  ParallelStableSortStability. Cancel/progress/reuse semantics for the
  ExecutionContext API and the ctx-aware for_each / parallel mergeSort
  mechanism tests they sit on top of.
- test/measurement_test.cpp: RayCast tests from manifold_test.cpp
  plus MinGap and TriangleDistance tests pulled out of
  properties_test.cpp. All "measurement-style queries on Manifolds":
  ray-mesh intersection, inter-manifold distance, the per-triangle
  distance primitive MinGap is built on.

Pure file moves: tests preserved verbatim, no behavior change.

Per the review on #1697:
- ForEachCtx and ParallelStableSort were briefly in their own
  parallel_test.cpp; merged into context_test.cpp on Emmett's
  suggestion that they're context-related and the file is too small
  to stand alone.
- ray_cast_test.cpp was a single-feature file; broadened and renamed
  to measurement_test.cpp with the MinGap and TriangleDistance tests
  that fit the same conceptual cluster.

Rebased onto upstream/master after #1694 merged. The two new
ExecutionContext tests from #1694
(ProgressReachesOneOnTrivialBooleans, SubBooleanProgress) plus the
Progress() == 1.0 cross-check assertions from #1694 land in
context_test.cpp; the ExecutionContextNoWorkNeeded test's expected
value updates accordingly.
