+++ test/CMakeLists.txt
@@ -25,6 +25,8 @@ set(
   polygon_test.cpp
   properties_test.cpp
   manifold_test.cpp
+  context_test.cpp
+  measurement_test.cpp
   boolean_test.cpp
   sdf_test.cpp
   smooth_test.cpp
